fix(front): move style files of event
