+++ client/src/pages/event.tsx
@@ -6,7 +6,7 @@ import EventOverview from "~/components/event/overview";
 import EventDescription from "~/components/event/description";
 import TweetButton from "~/components/button/tweet-button";
 import { client } from "~/libs/apiClient";
-import styles from "~/styles/event/event.module.scss";
+import styles from "~/styles/pages/event.module.scss";
 
 // 登壇枠の型
 export type SpeakerQuotaType = {
