+++ client/src/styles/components/event/date.module.scss
@@ -1,4 +1,4 @@
-@use "../variables" as var;
+@use "../../variables" as var;
 
 .enshita-event-date {
   color: var.$global-black;
