+++ client/src/styles/components/event/description.module.scss
@@ -1,4 +1,4 @@
-@use "../variables" as var;
+@use "../../variables" as var;
 
 .enshita-event-description-content {
   width: 100%;
