+++ client/src/styles/components/event/overview.module.scss
@@ -1,4 +1,4 @@
-@use "../variables" as var;
+@use "../../variables" as var;
 
 .enshita-event-overview {
   width: 100%;
